+++ CHANGELOG.md
@@ -16,4 +16,4 @@
 
 ## [0.2.1] - 2023-03-21
 
-- GPT-4 models supported (experimental)
+- GPT-4 models supported (in `normal` mode)